requirements.txt
cryptg==0.5.0.post0
psutil==6.1.0
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
rsa==4.9
Telethon==1.37.0
